continue laravel 21-25

+++ app/Http/Controllers/JobController.php
@@ -2,8 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\JobPosted;
 use App\Models\Job;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -42,16 +46,22 @@ class JobController extends Controller
             'salary' => ['required']
         ]);
 
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
 
+        Mail::to($job->employer->user)->queue(
+            new JobPosted($job)
+        );
+
         return redirect('/jobs');
     }
 
     public function edit(Job $job) {
+        //Gate::authorize('edit.job', $job);
+
         return view('jobs.edit', ['job' => $job]);
     }
 
@@ -72,6 +82,7 @@ class JobController extends Controller
     }
 
     public function destroy(Job $job) {
+        Gate::authorize('edit.job', $job);
         $job->delete();
 
         return redirect('/jobs');
